compute_ecmascript_module_exports
