+++ turbopack/crates/turbopack-ecmascript/src/lib.rs
@@ -117,6 +117,7 @@ use crate::{
         analyze_ecmascript_module,
         async_module::OptionAsyncModule,
         esm::{UrlRewriteBehavior, base::EsmAssetReferences, export},
+        exports::compute_ecmascript_module_exports,
     },
     side_effect_optimization::reference::EcmascriptModulePartReference,
     swc_comments::{CowComments, ImmutableComments},
@@ -584,7 +585,7 @@ impl EcmascriptAnalyzable for EcmascriptModuleAsset {
             async_module: analyze_ref.async_module,
             generate_source_map,
             original_source_map: analyze_ref.source_map,
-            exports: analyze_ref.exports,
+            exports: self.get_exports().to_resolved().await?,
             async_module_info,
         }
         .cell())
@@ -806,7 +807,7 @@ impl ChunkableModule for EcmascriptModuleAsset {
 impl EcmascriptChunkPlaceable for EcmascriptModuleAsset {
     #[turbo_tasks::function]
     async fn get_exports(self: Vc<Self>) -> Result<Vc<EcmascriptExports>> {
-        Ok(*self.analyze().await?.exports)
+        Ok(*compute_ecmascript_module_exports(self, None).await?.exports)
     }
 
     #[turbo_tasks::function]
