+++ turbopack/crates/turbopack-ecmascript/src/tree_shake/part/module.rs
@@ -20,7 +20,8 @@ use crate::{
     },
     parse::ParseResult,
     references::{
-        FollowExportsResult, analyze_ecmascript_module, esm::FoundExportType, follow_reexports,
+        FollowExportsResult, analyze_ecmascript_module, esm::FoundExportType,
+        exports::compute_ecmascript_module_exports, follow_reexports,
     },
     rename::module::EcmascriptModuleRenameModule,
     tree_shake::{
@@ -105,7 +106,7 @@ impl EcmascriptAnalyzable for EcmascriptModulePartAsset {
             async_module: analyze_ref.async_module,
             generate_source_map,
             original_source_map: analyze_ref.source_map,
-            exports: analyze_ref.exports,
+            exports: self.get_exports().to_resolved().await?,
             async_module_info,
         }
         .cell())
@@ -230,7 +231,7 @@ impl EcmascriptModulePartAsset {
     #[turbo_tasks::function]
     pub async fn is_async_module(self: Vc<Self>) -> Result<Vc<bool>> {
         let this = self.await?;
-        let result = analyze(*this.full_module, this.part.clone());
+        let result = analyze_ecmascript_module(*this.full_module, Some(this.part.clone()));
 
         if let Some(async_module) = *result.await?.async_module.await? {
             Ok(async_module.is_self_async(self.references()))
@@ -331,7 +332,7 @@ impl Module for EcmascriptModulePartAsset {
             return Ok(Vc::cell(references));
         }
 
-        let analyze = analyze(*self.full_module, self.part.clone());
+        let analyze = analyze_ecmascript_module(*self.full_module, Some(self.part.clone()));
 
         Ok(analyze.references())
     }
@@ -350,8 +351,12 @@ impl Module for EcmascriptModulePartAsset {
 #[turbo_tasks::value_impl]
 impl EcmascriptChunkPlaceable for EcmascriptModulePartAsset {
     #[turbo_tasks::function]
-    async fn get_exports(self: Vc<Self>) -> Result<Vc<EcmascriptExports>> {
-        Ok(*self.analyze().await?.exports)
+    async fn get_exports(&self) -> Result<Vc<EcmascriptExports>> {
+        Ok(
+            *compute_ecmascript_module_exports(*self.full_module, Some(self.part.clone()))
+                .await?
+                .exports,
+        )
     }
 
     #[turbo_tasks::function]
@@ -391,16 +396,8 @@ impl ChunkableModule for EcmascriptModulePartAsset {
 impl EcmascriptModulePartAsset {
     #[turbo_tasks::function]
     pub(super) fn analyze(&self) -> Vc<AnalyzeEcmascriptModuleResult> {
-        analyze(*self.full_module, self.part.clone())
+        analyze_ecmascript_module(*self.full_module, Some(self.part.clone()))
     }
-}
-
-#[turbo_tasks::function]
-fn analyze(
-    module: Vc<EcmascriptModuleAsset>,
-    part: ModulePart,
-) -> Vc<AnalyzeEcmascriptModuleResult> {
-    analyze_ecmascript_module(module, Some(part))
 }
 
 #[turbo_tasks::value_impl]
